+++ checks/checks.nix
@@ -32,7 +32,8 @@ in
 
   domain-empty = test testlib.domain domain/empty;
   domain-linux = test testlib.domain domain/template-linux;
-  domain-windows = test testlib.domain domain/template-windows;
+  domain-windows-1 = test testlib.domain domain/template-windows-1;
+  domain-windows-2 = test testlib.domain domain/template-windows-2;
   domain-win11 = test testlib.domain domain/win11;
 
   pool-empty = test testlib.pool pool/empty;