+++ templates/domain/linux.nix
@@ -4,13 +4,14 @@ stuff@{ packages, ... }:
 , memory ? { count = 4; unit = "GiB"; }
 , storage_vol
 , install_vol ? null
+, virtio_drive ? true
 , virtio_video ? true
 , ...
 }:
 let
   base = (import ./base.nix stuff).q35
     {
-      inherit name uuid memory storage_vol install_vol virtio_video;
+      inherit name uuid memory storage_vol install_vol virtio_drive virtio_video;
       virtio_net = true;
     };
 in